+++ internal/handler/web/main.go
@@ -4,13 +4,11 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/ucl-arc-tre/portal/internal/service/agreements"
 	"github.com/ucl-arc-tre/portal/internal/service/profile"
-	"github.com/ucl-arc-tre/portal/internal/service/training"
 )
 
 type Handler struct {
 	agreements *agreements.Service
 	profile    *profile.Service
-	training   *training.Service
 }
 
 func New() *Handler {
@@ -18,6 +16,5 @@ func New() *Handler {
 	return &Handler{
 		agreements: agreements.New(),
 		profile:    profile.New(),
-		training:   training.New(),
 	}
 }
