+++ internal/service/profile/certificate/nhsd.go
@@ -1,4 +1,4 @@
-package pdf
+package certificate
 
 import (
 	"encoding/base64"
