+++ internal/service/profile/certificate/types.go
@@ -1,4 +1,4 @@
-package pdf
+package certificate
 
 import (
 	"strings"
