+++ web/src/components/profile/approved-researcher/TrainingCertificate.module.css
@@ -1,0 +1,19 @@
+.wrapper {
+  max-width: 400px;
+  margin: 2rem auto;
+  padding: 1rem;
+}
+
+.text {
+  p {
+    padding-top: 0.7rem;
+  }
+
+  a {
+    color: #9f84f2;
+  }
+
+  a:hover {
+    text-decoration: underline;
+  }
+}
