--- web/src/components/profile/approved-researcher/TrainingCertificateForm.module.css
@@ -1,5 +1,0 @@
-.wrapper {
-  max-width: 400px;
-  margin: 2rem auto;
-  padding: 1rem;
-}
